refactor(Bibliography Renderer): Abstract collecting references into a separate function
